Almost have basic method for AST conversion down
I'm trying to parse the ANTLR AST into a more clojurey datastructure.
This approach is being taken through the visitor pattern with clojure's
multimethods. The issue I'm having right now is that I'm trying to use
reflection to map the inner classes of MiniJavaParser to clojure
keywords, so that I can use `:main-method-declartion` instead of
`MiniJavaParser$MainMethodDeclarationContext`. I have the member to
keyword process down, it's just a matter of accessing the different
context classes, which `reflect` is not giving me.